refactor(userInterface): Organize userInterface into separate files
