+++ src/userInterface/createImageUI.js
@@ -1,175 +1,13 @@
 import vtkColorMaps from 'vtk.js/Sources/Rendering/Core/ColorTransferFunction/ColorMaps';
-import vtkURLExtract from 'vtk.js/Sources/Common/Core/URLExtract';
 import vtkPiecewiseGaussianWidget from 'vtk.js/Sources/Interaction/Widgets/PiecewiseGaussianWidget';
 
-import processFiles from './processFiles';
+import getContrastSensitiveStyle from './getContrastSensitiveStyle';
 
 import style from './ItkVtkImageViewer.mcss';
 
-import logoIcon from './icons/logo.png';
-import toggleIcon from './icons/toggle.svg';
-import uploadIcon from './icons/upload.svg';
-import screenshotIcon from './icons/screenshot.svg';
 import shadowIcon from './icons/shadow.svg';
 import sampleDistanceIcon from './icons/sample-distance.svg';
 import gradientOpacityIcon from './icons/gradient.svg';
-
-function getContrastSensitiveStyle(cssClasses, isBackgroundDark) {
-  const stylePostFix = isBackgroundDark ? 'DarkBG' : 'BrightBG';
-  const contrastSensitiveStyle = {};
-  cssClasses.forEach((name) => {
-    contrastSensitiveStyle[name] = style[`${name}${stylePostFix}`];
-  });
-  return contrastSensitiveStyle;
-}
-
-function getRootContainer(container) {
-  const workContainer = document.querySelector('.content');
-  const rootBody = document.querySelector('body');
-
-  return container || workContainer || rootBody;
-}
-
-function preventDefaults(e) {
-  e.preventDefault();
-  e.stopPropagation();
-}
-
-function emptyContainer(container) {
-  if (container) {
-    while (container.firstChild) {
-      container.removeChild(container.firstChild);
-    }
-  }
-}
-
-function createFileDragAndDrop(container, onDataChange) {
-  const myContainer = getRootContainer(container);
-
-  const fileContainer = document.createElement('div');
-  fileContainer.innerHTML = `<div class="${
-    style.bigFileDrop
-  }"/><input type="file" class="file" style="display: none;" multiple/>`;
-  myContainer.appendChild(fileContainer);
-
-  const fileInput = fileContainer.querySelector('input');
-
-  function handleFile(e) {
-    preventDefaults(e);
-    const dataTransfer = e.dataTransfer;
-    const files = e.target.files || dataTransfer.files;
-    myContainer.removeChild(fileContainer);
-    const use2D = !!vtkURLExtract.extractURLParameters().use2D;
-    onDataChange(myContainer, { files, use2D });
-  }
-
-  fileInput.addEventListener('change', handleFile);
-  fileContainer.addEventListener('drop', handleFile);
-  fileContainer.addEventListener('click', (e) => fileInput.click());
-  fileContainer.addEventListener('dragover', preventDefaults);
-}
-
-function createLoadingProgress(container) {
-  const myContainer = getRootContainer(container);
-
-  const loading = document.createElement('div');
-  loading.setAttribute('class', style.loading);
-  myContainer.appendChild(loading);
-
-  const progressContainer = document.createElement('div');
-  progressContainer.setAttribute('class', style.progress);
-  myContainer.appendChild(progressContainer);
-
-  function progressCallback(progressEvent) {
-    const percent = Math.floor(
-      100 * progressEvent.loaded / progressEvent.total
-    );
-    progressContainer.innerHTML = `${percent}%`;
-  }
-
-  return progressCallback;
-}
-
-function addLogo(uiContainer) {
-  const logo = new Image();
-  logo.src = logoIcon;
-  logo.setAttribute('class', style.logo);
-  uiContainer.appendChild(logo);
-}
-
-function createMainUI(rootContainer, isBackgroundDark, imageSource, view) {
-  const uiContainer = document.createElement('div');
-  rootContainer.appendChild(uiContainer);
-  uiContainer.setAttribute('class', style.uiContainer);
-
-  const contrastSensitiveStyle = getContrastSensitiveStyle(
-    ['uiToggleButton', 'uploadButton', 'screenshotButton'],
-    isBackgroundDark
-  );
-
-  const mainUIGroup = document.createElement('div');
-  mainUIGroup.setAttribute('class', style.uiGroup);
-
-  const mainUIRow = document.createElement('div');
-  mainUIRow.setAttribute('class', style.mainUIRow);
-  mainUIRow.className += ' js-toggle';
-  mainUIGroup.appendChild(mainUIRow);
-
-  function toggleUIVisibility() {
-    const elements = uiContainer.querySelectorAll('.js-toggle');
-    let count = elements.length;
-    const toggleElementStyle = window.getComputedStyle(elements[0]);
-    const expanded = toggleElementStyle.getPropertyValue('display') === 'flex';
-    if (!expanded) {
-      while (count--) {
-        elements[count].style.display = 'flex';
-      }
-    } else {
-      while (count--) {
-        elements[count].style.display = 'none';
-      }
-    }
-  }
-  const uiToggleButton = document.createElement('div');
-  uiToggleButton.innerHTML = `<div class="${
-    contrastSensitiveStyle.uiToggleButton
-  }">${toggleIcon}</div>`;
-  uiToggleButton.addEventListener('click', toggleUIVisibility);
-  uiContainer.appendChild(uiToggleButton);
-
-  const uploadButton = document.createElement('div');
-  uploadButton.innerHTML = `<div class="${
-    contrastSensitiveStyle.uploadButton
-  }">${uploadIcon}</div><input type="file" class="file" style="display: none;" multiple/>`;
-  const fileInput = uploadButton.querySelector('input');
-
-  function handleFile(e) {
-    preventDefaults(e);
-    const dataTransfer = e.dataTransfer;
-    const files = e.target.files || dataTransfer.files;
-    processFiles(rootContainer, { files });
-  }
-
-  fileInput.addEventListener('change', handleFile);
-  uploadButton.addEventListener('drop', handleFile);
-  uploadButton.addEventListener('click', (e) => fileInput.click());
-  uploadButton.addEventListener('dragover', preventDefaults);
-  mainUIRow.appendChild(uploadButton);
-
-  const screenshotButton = document.createElement('div');
-  screenshotButton.innerHTML = `<div class="${
-    contrastSensitiveStyle.screenshotButton
-  }">${screenshotIcon}</div>`;
-  function takeScreenshot() {
-    view.openCaptureImage();
-  }
-  screenshotButton.addEventListener('click', takeScreenshot);
-  mainUIRow.appendChild(screenshotButton);
-
-  uiContainer.appendChild(mainUIGroup);
-
-  return uiContainer;
-}
 
 function createUseShadowToggle(
   uiContainer,
@@ -319,7 +157,7 @@ function createSampleDistanceSlider(
     <div class="${contrastSensitiveStyle.sampleDistanceButton}">
       ${sampleDistanceIcon}
     </div>
-    <input type="range" min="0" max="1" value="0.4" step="0.01"
+    <input type="range" min="0" max="1" value="0.3" step="0.01"
       class="${style.slider} js-spacing" />`;
   const spacingElement = sliderEntry.querySelector('.js-spacing');
   function updateSpacing() {
@@ -414,12 +252,4 @@ function createImageUI(
   uiContainer.appendChild(imageUIGroup);
 }
 
-export default {
-  addLogo,
-  createFileDragAndDrop,
-  createLoadingProgress,
-  createMainUI,
-  createImageUI,
-  emptyContainer,
-  getRootContainer,
-};
+export default createImageUI;
