+++ src/userInterface/emptyContainer.js
@@ -1,0 +1,9 @@
+function emptyContainer(container) {
+  if (container) {
+    while (container.firstChild) {
+      container.removeChild(container.firstChild);
+    }
+  }
+}
+
+export default emptyContainer;
