+++ src/userInterface/getRootContainer.js
@@ -1,0 +1,8 @@
+function getRootContainer(container) {
+  const workContainer = document.querySelector('.content');
+  const rootBody = document.querySelector('body');
+
+  return container || workContainer || rootBody;
+}
+
+export default getRootContainer;
