+++ src/userInterface/icons/logo.png
