+++ src/userInterface/index.js
@@ -1,0 +1,17 @@
+import addLogo from './addLogo';
+import createFileDragAndDrop from './createFileDragAndDrop';
+import createLoadingProgress from './createLoadingProgress';
+import createMainUI from './createMainUI';
+import createImageUI from './createImageUI';
+import emptyContainer from './emptyContainer';
+import getRootContainer from './getRootContainer';
+
+export default {
+  addLogo,
+  createFileDragAndDrop,
+  createLoadingProgress,
+  createMainUI,
+  createImageUI,
+  emptyContainer,
+  getRootContainer,
+};
