+++ src/userInterface/preventDefaults.js
@@ -1,0 +1,6 @@
+function preventDefaults(e) {
+  e.preventDefault();
+  e.stopPropagation();
+}
+
+export default preventDefaults;
